refactor(engine): extract ProgressionProcessor + WeeklyFinancesProcessor (Phase 2 PR-6)
Verbatim moves of two more §2 rows out of shared/engine/game-engine.ts,
following the PR-5 processor pattern (stateless class + WeekContext).

ProgressionProcessor: checkProgressionGates, updateAccessTiers,
checkProducerTierUnlocks, checkCampaignCompletion.

WeeklyFinancesProcessor: calculateWeeklyBurn, calculateWeeklyBurnWithBreakdown,
calculateWeeklyFinancials, generateFinancialBreakdown, generateEconomicInsights.

Every method keeps a thin same-signature delegate on GameEngine so no caller
outside the file changes (external callers: updateAccessTiers via task-0001 test,
calculateWeeklyFinancials via scripts/debug-tests). advanceWeek pipeline call
order is unchanged; no RNG-order change. The single money-update point
(this.gameState.money = finalMoney in advanceWeek) is untouched.

WeekContext gains a financialSystem field so the finances processor delegates
to the engine's FinancialSystem exactly as before.

game-engine.ts 4861->4637. Golden master byte-identical (zero snapshot updates);
full suite 652 green.
